refactor: split routers into modules
